requirements.txt
tqdm
yara-python
python-dotenv
filetype
PyYAML
psutil
watchdog
python-dateutil
vt-py==0.17.4
cryptography
frida
frida-tools
pynput
requests

flask
mpld3
numpy
matplotlib

dash
plotly
dash-core-components 
dash-html-components
psutil
dash_bootstrap_components
pandas

clamd